\name{definition}
\alias{definition}
\title{
Add rectangles on Hilbert curve

}
\description{
Add rectangles on Hilbert curve

}
\usage{
definition(object, ir, gp = gpar(fill = "red"), mean_mode = c("w0", "absolute", "weighted"))}
\arguments{

  \item{object}{-object}
  \item{ir}{-ir}
  \item{value}{-value}
  \item{fill}{-fill}
  \item{mean_mode}{-mean_mode}
}
\details{
only color/fill can be mapped to cicles}
