\name{hc_layer-dispatch}
\alias{hc_layer}
\title{
Method dispatch page for hc_layer
}
\description{
Method dispatch page for \code{hc_layer}.
}
\section{Dispatch}{
\code{hc_layer} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_layer,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
\item \code{\link{hc_layer,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
