\name{hc_layer-HilbertCurve-method}
\alias{hc_layer,HilbertCurve-method}
\alias{hc_layer}
\title{
hc_layer

}
\description{
hc_layer

}
\usage{
\S4method{hc_layer}{HilbertCurve}(object, ir, col = "red", mean_mode = c("w0", "absolute", "weighted"))}
\arguments{

  \item{object}{-object}
  \item{ir}{-ir}
  \item{col}{-col}
  \item{mean_mode}{-mean_mode}
}
