\name{hc_points-dispatch}
\alias{hc_points}
\title{
Method dispatch page for hc_points
}
\description{
Method dispatch page for \code{hc_points}.
}
\section{Dispatch}{
\code{hc_points} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_points,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
\item \code{\link{hc_points,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
