\name{hc_polygon-dispatch}
\alias{hc_polygon}
\title{
Method dispatch page for hc_polygon
}
\description{
Method dispatch page for \code{hc_polygon}.
}
\section{Dispatch}{
\code{hc_polygon} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_polygon,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
\item \code{\link{hc_polygon,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
