\name{hc_rect-dispatch}
\alias{hc_rect}
\title{
Method dispatch page for hc_rect
}
\description{
Method dispatch page for \code{hc_rect}.
}
\section{Dispatch}{
\code{hc_rect} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_rect,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
\item \code{\link{hc_rect,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
