\name{hc_save-HilbertCurve-method}
\alias{hc_save,HilbertCurve-method}
\alias{hc_save}
\title{
Save Hilbert curve as PNG figure

}
\description{
Save Hilbert curve as PNG figure

}
\usage{
\S4method{hc_save}{HilbertCurve}(object, file = "Rplot.png", grid = 0)}
\arguments{

  \item{object}{A \code{\link{HilbertCurve-class}} object}
  \item{file}{file name}
  \item{grid}{add grid lines, should be an integer number}
}
\value{
No value is returned

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
\examples{
# see hc_layer
NULL

}
