\name{hc_segmented_points-HilbertCurve-method}
\alias{hc_segmented_points,HilbertCurve-method}
\title{
hc_segmented_points

}
\description{
hc_segmented_points

}
\usage{
\S4method{hc_segmented_points}{HilbertCurve}(object, ir, gp = gpar(), np = max(c(2, 10 - hc_level(object))),
    mean_mode = c("w0", "absolute", "weighted"), shape = c("circle", "square", "triangle", "hexagon", "star"))}
\arguments{

  \item{object}{-object}
  \item{ir}{-ir}
  \item{gp}{-gp}
  \item{np}{-np}
  \item{mean_mode}{-mean_mode}
  \item{shape}{-shape}
}
\details{
only color/fill can be mapped to cicles

}
\alias{hc_segmented_points}
