\name{hc_segments-dispatch}
\alias{hc_segments}
\title{
Method dispatch page for hc_segments
}
\description{
Method dispatch page for \code{hc_segments}.
}
\section{Dispatch}{
\code{hc_segments} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_segments,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
\item \code{\link{hc_segments,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
