\name{hc_text-dispatch}
\alias{hc_text}
\title{
Method dispatch page for hc_text
}
\description{
Method dispatch page for \code{hc_text}.
}
\section{Dispatch}{
\code{hc_text} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_text,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
\item \code{\link{hc_text,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
