\name{hc_which-dispatch}
\alias{hc_which}
\title{
Method dispatch page for hc_which
}
\description{
Method dispatch page for \code{hc_which}.
}
\section{Dispatch}{
\code{hc_which} can be dispatched on following classes:

\itemize{
\item \code{\link{hc_which,HilbertCurve-method}}, \code{\link{HilbertCurve-class}} class method
\item \code{\link{hc_which,GenomicHilbertCurve-method}}, \code{\link{GenomicHilbertCurve-class}} class method
}
}
\examples{
# no example
NULL


}
