\name{is_white}
\alias{is_white}
\title{
Whether the color is white
}
\description{
Whether the color is white
}
\usage{
is_white(r, g, b, maxColorValue = 1)
}
\arguments{

  \item{r}{Red channel.}
  \item{g}{Green channel.}
  \item{b}{Blue channel.}
  \item{maxColorValue}{1 or 255.}

}
\examples{
# There is no example
NULL
}
