\name{show-HilbertCurve-method}
\alias{show,HilbertCurve-method}
\alias{show}
\title{
Print the HilbertCurve object
}
\description{
Print the HilbertCurve object
}
\usage{
\S4method{show}{HilbertCurve}(object)
}
\arguments{

  \item{object}{A \code{\link{HilbertCurve-class}} object.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
hc = HilbertCurve(1, 100)
 hc
}
