\name{hc_normal_points-HilbertCurve-method}
\alias{hc_normal_points,HilbertCurve-method}
\title{
hc_normal_points

}
\description{
hc_normal_points

}
\usage{
\S4method{hc_normal_points}{HilbertCurve}(object, ir, gp = gpar(), size = unit(1, "char"))}
\arguments{

  \item{object}{-object}
  \item{ir}{-ir}
  \item{gp}{-gp}
  \item{size}{-size}
}
\alias{hc_normal_points}
