\name{hc_rect-HilbertCurve-method}
\alias{hc_rect,HilbertCurve-method}
\title{
Add rectangles on Hilbert curve

}
\description{
Add rectangles on Hilbert curve

}
\usage{
\S4method{hc_rect}{HilbertCurve}(object, ir, gp = gpar(fill = "red"), mean_mode = c("w0", "absolute", "weighted"))}
\arguments{

  \item{object}{-object}
  \item{ir}{-ir}
  \item{value}{-value}
  \item{fill}{-fill}
  \item{mean_mode}{-mean_mode}
}
\alias{hc_rect}
