\name{hc_segments-HilbertCurve-method}
\alias{hc_segments,HilbertCurve-method}
\alias{hc_segments}
\title{
Add line segments to Hilbert curve

}
\description{
Add line segments to Hilbert curve

}
\usage{
\S4method{hc_segments}{HilbertCurve}(object, ir, gp = gpar())}
\arguments{

  \item{object}{-object}
  \item{ir}{a \code{\link[IRanges]{IRanges}} object in which each interval corresponds to a segment in the curve.}
  \item{gp}{grahical parameters for lines}
}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
