\name{unzoom-HilbertCurve-method}
\alias{unzoom,HilbertCurve-method}
\alias{unzoom}
\title{
Transform zoomed positions to their original values

}
\description{
Transform zoomed positions to their original values

}
\usage{
\S4method{unzoom}{HilbertCurve}(object, x)}
\arguments{

  \item{object}{A \code{\link{HilbertCurve-class}} object}
  \item{x}{positions}
}
\details{
The function is used internally

}
\value{
Original positions

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
