\name{zoom-HilbertCurve-method}
\alias{zoom,HilbertCurve-method}
\alias{zoom}
\title{
Zoom original positions

}
\description{
Zoom original positions

}
\usage{
\S4method{zoom}{HilbertCurve}(object, x)}
\arguments{

  \item{object}{A \code{\link{HilbertCurve-class}} object}
  \item{x}{positions}
}
\details{
The function is used internally

}
\value{
Zoomed positions

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
